+++ src/Service/Lambda/src/Exception/ENILimitReachedException.php
@@ -6,8 +6,10 @@ use AsyncAws\Core\Exception\Http\ServerException;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 /**
- * Lambda was not able to create an elastic network interface in the VPC, specified as part of Lambda function
- * configuration, because the limit for network interfaces has been reached.
+ * Lambda couldn't create an elastic network interface in the VPC, specified as part of Lambda function configuration,
+ * because the limit for network interfaces has been reached. For more information, see Lambda quotas.
+ *
+ * @see https://docs.aws.amazon.com/lambda/latest/dg/gettingstarted-limits.html
  */
 final class ENILimitReachedException extends ServerException
 {