+++ src/Service/Lambda/src/Exception/InvalidSubnetIDException.php
@@ -6,7 +6,7 @@ use AsyncAws\Core\Exception\Http\ServerException;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 /**
- * The Subnet ID provided in the Lambda function VPC configuration is invalid.
+ * The subnet ID provided in the Lambda function VPC configuration is not valid.
  */
 final class InvalidSubnetIDException extends ServerException
 {