+++ src/Service/Lambda/src/Exception/PolicyLengthExceededException.php
@@ -6,9 +6,9 @@ use AsyncAws\Core\Exception\Http\ClientException;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 /**
- * The permissions policy for the resource is too large. Learn more.
+ * The permissions policy for the resource is too large. For more information, see Lambda quotas.
  *
- * @see https://docs.aws.amazon.com/lambda/latest/dg/limits.html
+ * @see https://docs.aws.amazon.com/lambda/latest/dg/gettingstarted-limits.html
  */
 final class PolicyLengthExceededException extends ClientException
 {